+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## v1.0.1 - 2022/01/04
 - [x] Add google drive backup
 - [x] Change path info and option show info
+- [x] Add install-wordpress
 
 ## v1.0.0 - 2022/01/01